+++ pkg/clientset/versioned/clientset.go
@@ -32,6 +32,7 @@ import (
 	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/pkg/clientset/versioned/typed/infrastructure/v1alpha2"
 	providerv1alpha1 "github.com/giantswarm/apiextensions/pkg/clientset/versioned/typed/provider/v1alpha1"
 	releasev1alpha1 "github.com/giantswarm/apiextensions/pkg/clientset/versioned/typed/release/v1alpha1"
+	toolingv1alpha1 "github.com/giantswarm/apiextensions/pkg/clientset/versioned/typed/tooling/v1alpha1"
 )
 
 type Interface interface {
@@ -43,6 +44,7 @@ type Interface interface {
 	InfrastructureV1alpha2() infrastructurev1alpha2.InfrastructureV1alpha2Interface
 	ProviderV1alpha1() providerv1alpha1.ProviderV1alpha1Interface
 	ReleaseV1alpha1() releasev1alpha1.ReleaseV1alpha1Interface
+	ToolingV1alpha1() toolingv1alpha1.ToolingV1alpha1Interface
 }
 
 // Clientset contains the clients for groups. Each group has exactly one
@@ -56,6 +58,7 @@ type Clientset struct {
 	infrastructureV1alpha2 *infrastructurev1alpha2.InfrastructureV1alpha2Client
 	providerV1alpha1       *providerv1alpha1.ProviderV1alpha1Client
 	releaseV1alpha1        *releasev1alpha1.ReleaseV1alpha1Client
+	toolingV1alpha1        *toolingv1alpha1.ToolingV1alpha1Client
 }
 
 // ApplicationV1alpha1 retrieves the ApplicationV1alpha1Client
@@ -91,6 +94,11 @@ func (c *Clientset) ProviderV1alpha1() providerv1alpha1.ProviderV1alpha1Interfac
 // ReleaseV1alpha1 retrieves the ReleaseV1alpha1Client
 func (c *Clientset) ReleaseV1alpha1() releasev1alpha1.ReleaseV1alpha1Interface {
 	return c.releaseV1alpha1
+}
+
+// ToolingV1alpha1 retrieves the ToolingV1alpha1Client
+func (c *Clientset) ToolingV1alpha1() toolingv1alpha1.ToolingV1alpha1Interface {
+	return c.toolingV1alpha1
 }
 
 // Discovery retrieves the DiscoveryClient
@@ -142,6 +150,10 @@ func NewForConfig(c *rest.Config) (*Clientset, error) {
 	if err != nil {
 		return nil, err
 	}
+	cs.toolingV1alpha1, err = toolingv1alpha1.NewForConfig(&configShallowCopy)
+	if err != nil {
+		return nil, err
+	}
 
 	cs.DiscoveryClient, err = discovery.NewDiscoveryClientForConfig(&configShallowCopy)
 	if err != nil {
@@ -161,6 +173,7 @@ func NewForConfigOrDie(c *rest.Config) *Clientset {
 	cs.infrastructureV1alpha2 = infrastructurev1alpha2.NewForConfigOrDie(c)
 	cs.providerV1alpha1 = providerv1alpha1.NewForConfigOrDie(c)
 	cs.releaseV1alpha1 = releasev1alpha1.NewForConfigOrDie(c)
+	cs.toolingV1alpha1 = toolingv1alpha1.NewForConfigOrDie(c)
 
 	cs.DiscoveryClient = discovery.NewDiscoveryClientForConfigOrDie(c)
 	return &cs
@@ -176,6 +189,7 @@ func New(c rest.Interface) *Clientset {
 	cs.infrastructureV1alpha2 = infrastructurev1alpha2.New(c)
 	cs.providerV1alpha1 = providerv1alpha1.New(c)
 	cs.releaseV1alpha1 = releasev1alpha1.New(c)
+	cs.toolingV1alpha1 = toolingv1alpha1.New(c)
 
 	cs.DiscoveryClient = discovery.NewDiscoveryClient(c)
 	return &cs