+++ pkg/clientset/versioned/fake/clientset_generated.go
@@ -40,6 +40,8 @@ import (
 	fakeproviderv1alpha1 "github.com/giantswarm/apiextensions/pkg/clientset/versioned/typed/provider/v1alpha1/fake"
 	releasev1alpha1 "github.com/giantswarm/apiextensions/pkg/clientset/versioned/typed/release/v1alpha1"
 	fakereleasev1alpha1 "github.com/giantswarm/apiextensions/pkg/clientset/versioned/typed/release/v1alpha1/fake"
+	toolingv1alpha1 "github.com/giantswarm/apiextensions/pkg/clientset/versioned/typed/tooling/v1alpha1"
+	faketoolingv1alpha1 "github.com/giantswarm/apiextensions/pkg/clientset/versioned/typed/tooling/v1alpha1/fake"
 )
 
 // NewSimpleClientset returns a clientset that will respond with the provided objects.
@@ -122,4 +124,9 @@ func (c *Clientset) ProviderV1alpha1() providerv1alpha1.ProviderV1alpha1Interfac
 // ReleaseV1alpha1 retrieves the ReleaseV1alpha1Client
 func (c *Clientset) ReleaseV1alpha1() releasev1alpha1.ReleaseV1alpha1Interface {
 	return &fakereleasev1alpha1.FakeReleaseV1alpha1{Fake: &c.Fake}
+}
+
+// ToolingV1alpha1 retrieves the ToolingV1alpha1Client
+func (c *Clientset) ToolingV1alpha1() toolingv1alpha1.ToolingV1alpha1Interface {
+	return &faketoolingv1alpha1.FakeToolingV1alpha1{Fake: &c.Fake}
 }